Give the runner a typed ResultFile and its own benchmark catalog
Three grok-ability problems in src/runner, one commit each would touch
the same lines, so together:

- The benchmark catalog -- the thing most edits want -- was buried at
  line 86 of bench-info.ts, inside the selection wizard. It now lives in
  benchmarks.ts, and the runner's BenchmarkInfo extends the results
  app's (the recorded shape) instead of duplicating it.

- Everything writing the result JSON was free functions over `any` in
  results.ts, connected to the app that reads the files only by
  convention -- and the convention had drifted: the app's types claimed
  a `url` field the runner never writes (it writes `app` + `query`),
  whatsBetter: "smaller" wasn't representable, and TIMEOUT / FRAMEWORK /
  BENCH_NAME args weren't either. The writer is now a ResultFile class
  in result-file.ts, typed against the app's ResultSet shape, and the
  app's types now describe what is actually in the files. Timing is
  exported from the shared types instead of being declared twice, and
  the runner's MarkEntry moved next to the file it's persisted in (as
  RecordedMark).

- arg.ts printed its flag table as an import side effect, so `pnpm copy`
  printed bench flags and add.ts had to avoid the import entirely. The
  print is now an explicit printFlagTable() called by `pnpm bench`.

The runner was never typechecked (CI only ran eslint/prettier), and
tsc --noEmit failed on main with four pre-existing errors. Those are
fixed (killable gets a declaration, so serve() returns a typed
KillableServer and index.ts drops its @ts-expect-error), and a root
lint:types script now runs under `pnpm lint`, which CI already calls.

Verified with a real run: --framework=ember --bench='1 item, 1k
updates' --count=2 --skip-build --headless writes a file
byte-compatible with the old shape (then deleted -- results come from
the benchmark machine); copy.ts and bench:add still reach their
prompts.
